requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.9.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.2
attrs==23.1.0
autopep8==1.6.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.9
boto3==1.21.23
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.1.7
django-allauth==0.47.0
django-api-gateway==0.0.7
django-ckeditor==6.2.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-debug-toolbar==4.0.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-paypal==2.0
django-rest-knox==4.1.0
django-storages==1.12.3
django-templated-mail==1.1.1
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.7
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.0.3
fqdn==1.5.1
geojson==2.5.0
gunicorn==20.1.0
h11==0.14.0
heroku==0.1.4
hvac==0.11.2
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.1.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
jwt==1.3.1
knox==0.1.14
langcodes==3.3.0
lazy-object-proxy==1.7.1
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.5
MouseInfo==0.1.3
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.1.1
outcome==1.2.0
packaging==23.1
pandocfilters==1.5.0
paramiko==2.10.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.0.1
Pint==0.21
platformdirs==3.5.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
PyAutoGUI==0.9.53
pybind11==2.10.4
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.7
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pymongo==3.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyowm==3.3.0
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-json-logger==2.0.7
python-vlc==3.0.16120
python3-openid==3.2.0
pytweening==1.0.7
pytz==2021.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.5.2
selenium==4.9.0
Send2Trash==1.8.2
shortuuid==1.0.8
six==1.14.0
smart-open==6.3.0
sniffio==1.3.0
social-auth-app-django==4.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.2.19
sqlparse==0.4.2
srsly==2.4.6
stack-data==0.6.2
static3==0.7.0
terminado==0.17.1
thinc==8.1.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2021.5
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.9
validators==0.18.2
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
whitenoise==6.0.0
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wrapt==1.13.3
wsproto==1.2.0
zipp==3.7.0
